refactor: migrated to newly created utils
